requirements.txt
impacket==0.10.0
python-nmap==0.6.1
loguru==0.5.3
scapy==2.4.4

